:pencil: fix id type

+++ src/main/java/com/fixingsolutions/budget/controller/BudgetController.java
@@ -36,7 +36,7 @@ public class BudgetController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable Integer id) {
         budgetService.delete(id);
     }
 }
