+++ src/main/java/com/fixingsolutions/budget/dtos/request/UpdateBudgetRequest.java
@@ -11,10 +11,10 @@ import java.util.List;
 @NoArgsConstructor
 @Data
 public class UpdateBudgetRequest {
-    private Long id;
+    private Integer id;
     private BigDecimal value;
     private Integer expectedHours;
     private Boolean approved;
-    private Long customerId;
-    private List<Long> serviceTypeIds;
+    private Integer customerId;
+    private List<Integer> serviceTypeIds;
 }
