+++ src/main/java/com/fixingsolutions/serviceType/controller/ServiceTypeController.java
@@ -33,7 +33,7 @@ public class ServiceTypeController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable Long id) throws Exception {
+    public void delete(@PathVariable Integer id) throws Exception {
         serviceTypeService.delete(id);
     }
